Store lastBuildOnBranch in BuildContext along with selectedBuild
